eliminate a warning
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/module-pgsql/trunk@4803 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ src/QorePGConnection.cc
@@ -965,7 +965,7 @@ int QorePGResult::add(const AbstractQoreNode *v, ExceptionSink *xsink) {
    if (ntype == NT_HASH) {
       const QoreHashNode *vh = reinterpret_cast<const QoreHashNode *>(v);
       Oid type = check_hash_type(vh, xsink);
-      if (type < 0)
+      if ((int)type < 0)
 	 return -1;
       const AbstractQoreNode *t = vh->getKeyValue("^value^");
       if (is_nothing(t) || is_null(t)) {
